+++ src/main/java/de/vette/idea/neos/lang/fusion/psi/FusionElementFactory.java
@@ -1,0 +1,33 @@
+package de.vette.idea.neos.lang.fusion.psi;
+
+import com.intellij.openapi.project.Project;
+import com.intellij.psi.PsiFileFactory;
+import de.vette.idea.neos.lang.fusion.FusionFileType;
+import org.jetbrains.annotations.NonNls;
+import org.jetbrains.annotations.NotNull;
+
+public class FusionElementFactory {
+    @NotNull
+    public static FusionPrototypeSignature createPrototypeSignature(@NotNull Project project, @NonNls @NotNull String name) {
+        final FusionFile dummyFile = createFusionFile(project, getPrototypeSignatureText(name));
+        return (FusionPrototypeSignature) dummyFile.getFirstChild().getFirstChild().getFirstChild();
+    }
+
+    @NotNull
+    public static FusionType createType(@NotNull Project project, @NonNls @NotNull String name) {
+        final FusionFile dummyFile = createFusionFile(project, getPrototypeSignatureText(name));
+        FusionPrototypeSignature signature = (FusionPrototypeSignature) dummyFile.getFirstChild().getFirstChild().getFirstChild();
+        return signature.getType();
+    }
+
+    @NotNull
+    public static FusionFile createFusionFile(@NotNull Project project, @NonNls @NotNull String text) {
+        @NonNls String filename = "dummy." + FusionFileType.INSTANCE.getDefaultExtension();
+        return (FusionFile)PsiFileFactory.getInstance(project)
+                .createFileFromText(filename, FusionFileType.INSTANCE, text);
+    }
+
+    protected static String getPrototypeSignatureText(String name) {
+        return "prototype(" + name + ") {}";
+    }
+}