+++ src/main/java/de/vette/idea/neos/lang/fusion/refactoring/FusionNamesValidator.java
@@ -1,0 +1,17 @@
+package de.vette.idea.neos.lang.fusion.refactoring;
+
+import com.intellij.lang.refactoring.NamesValidator;
+import com.intellij.openapi.project.Project;
+import org.jetbrains.annotations.NotNull;
+
+class FusionNamesValidator implements NamesValidator {
+    @Override
+    public boolean isKeyword(@NotNull String name, Project project) {
+        return false;
+    }
+
+    @Override
+    public boolean isIdentifier(@NotNull String name, Project project) {
+        return true;
+    }
+}