+++ CHANGELOG.md
@@ -3,11 +3,14 @@
 # IntelliJ Neos Changelog
 
 ## [Unreleased]
+### Added
+- Rename refactoring for Fusion prototypes
+- Find usages of Fusion prototypes
+
 ### Fixed
 - Index split settings files (Settings.*.yaml)
 - Weird behavior when trying to navigate to prototype definitions (sometimes only the namespace was clickable)
 
-### Security
 ## [1.5.0]
 ### Added